CourseAnalyticsExport inherits from organizer, reorganize current Analytics::Export implementation so we can add support modules, rename Analytics::Export to Analytics::Export::Model, better organization for Export::Builder
