+++ app/analytics_exports/course_event_export.rb
@@ -1,5 +1,5 @@
 class CourseEventExport
-  include Analytics::Export
+  include Analytics::Export::Model
 
   rows_by :events
