+++ app/analytics_exports/course_predictor_export.rb
@@ -1,5 +1,5 @@
 class CoursePredictorExport
-  include Analytics::Export
+  include Analytics::Export::Model
 
   rows_by :events
