+++ app/analytics_exports/course_user_aggregate_export.rb
@@ -1,5 +1,5 @@
 class CourseUserAggregateExport
-  include Analytics::Export
+  include Analytics::Export::Model
 
   rows_by :users
